+++ include/daemon/WriteOnlyDaemonBase.hpp
@@ -14,21 +14,21 @@ public:
     virtual ~WriteOnlyDaemonBase() = default;
 
     virtual auto writeTxToBlockchain(std::string&& txid_input,
-                                     std::size_t index,
+                                     std::int64_t index,
                                      std::vector<std::byte>&& metadata,
-                                     std::size_t burn_value,
+                                     std::int64_t burn_value,
                                      std::vector<
                                          std::pair<std::string,
-                                                   std::size_t>>&& outputs) const
+                                                   std::int64_t>>&& outputs) const
         -> utilxx::Result<void, DaemonError>;
 
     virtual auto generateRawTx(std::string&& input_txid,
-                               std::size_t index,
+                               std::int64_t index,
                                std::vector<std::byte>&& metadata,
-                               std::size_t burn_value,
+                               std::int64_t burn_value,
                                std::vector<
                                    std::pair<std::string,
-                                             std::size_t>>&& outputs) const
+                                             std::int64_t>>&& outputs) const
         -> utilxx::Result<std::vector<std::byte>,
                           DaemonError> = 0;
 
@@ -48,7 +48,7 @@ public:
 auto make_writing_daemon(const std::string& host,
                          const std::string& user,
                          const std::string& password,
-                         std::size_t port,
+                         std::int64_t port,
                          core::Coin coin)
     -> std::unique_ptr<WriteOnlyDaemonBase>;
