+++ src/core/OwnershipTransferOp.cpp
@@ -7,8 +7,8 @@ using buddy::core::OwnershipTransferOp;
 OwnershipTransferOp::OwnershipTransferOp(Entry&& entry,
                                          std::string&& old_owner,
                                          std::string&& new_owner,
-                                         std::size_t block,
-                                         std::size_t value)
+                                         std::int64_t block,
+                                         std::int64_t value)
     : entry_(std::move(entry)),
       old_owner_(std::move(old_owner)),
       new_owner_(std::move(new_owner)),
@@ -76,13 +76,13 @@ auto OwnershipTransferOp::getOwner()
 }
 
 auto OwnershipTransferOp::getBlock() const
-    -> std::size_t
+    -> std::int64_t
 {
     return block_;
 }
 
 auto OwnershipTransferOp::getValue() const
-    -> std::size_t
+    -> std::int64_t
 {
     return value_;
 }
